requirements.txt
colorama==0.4.6
emoji==2.12.1
langchain==0.2.11
langchain_community==0.2.10
matplotlib==3.8.4
ollama==0.3.1
pandas==1.5.3
pandasai==2.2.12
psycopg2_binary==2.9.9
streamlit==1.36.0
streamlit_extras==0.4.3
streamlit_multipage==0.0.18
tqdm==4.66.4
